Crate embed_wasm

This crate provides utilities for serving 'cargo-web' build output (rust compiled as wasm and associated html/css/etc files) included in native binaries as HTTP responses

Designed for use with the embed-wasm-build crate. See embed-wasm-example for a full example.

Macros

include_wasm
Imports the generated static wasm blobs from the build output directory as generated by 'embed_wasm_build::compile_wasm'

Structs

StaticLookup
Wraps a hashmap of static content generated at compile time and provides convenience functions for resolving static content as 'hyper::Response' when given a path.

Enums

IndexHandling
Enum controlling how requests with path '/' are handled